Initial attempt to add ViewModels foiled by type-erasure
Because instantiating a templated class for getting its class type won't work. I was in a hurry or I would've stopped before changing ~20 files. We can't use templates to wrap ViewModels because the types get erased and then we can't use the types to access the key with the ViewModelProviders.